playground/requirements.txt
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.20.0
requests>=2.25.0
python-dotenv>=0.19.0